requirements.txt
aioauth-client==0.17.3
fastapi==0.44.0
gunicorn==20.0.4
pydantic[email]
python-dotenv==0.10.3
uvicorn==0.10.8
